+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.5.0 (2020-07-15)
+Added option for users to change Spider value in Content API calls (defaults to enabled)
+Renamed Global Vars to Custom Fields and updated helper text on Content Settings page
+
 ## v3.4.4 (2020-07-01)
 Fixed bug where “Scout from Sailthru” page was created for all sites, rather than just Scout-enabled sites.